refactor: edit workflows-editor.context
- add separated context
- use component provider without a function
