+++ frontend/src/context/workflows/workflows-editor.context/workflow-edges.context.tsx
@@ -1,0 +1,61 @@
+import React, { useCallback, useEffect, useState } from "react";
+import { Edge } from "reactflow";
+import localForage from "services/config/local-forage.config";
+import { createCustomContext } from "utils";
+
+export interface IWorkflowsEdgesContext {
+  edges: Edge[]
+  setEdges: React.Dispatch<React.SetStateAction<Edge[]>>
+  fetchForageWorkflowEdges: () => Promise<Edge[]>
+}
+
+export const [WorkflowsEdgesContext, useWorkflowsEdges] =
+  createCustomContext<IWorkflowsEdgesContext>('WorkflowsEdges Context')
+
+const WorkflowsEdgesProvider: React.FC<{children:React.ReactNode}> = ({ children }) => {
+  const [edges, setEdges] = useState<Edge[]>([])
+  const [loadingEdges, setLoadingEdges] = useState<boolean>(true)
+  
+  const setForageWorkflowEdges = useCallback(async (edges: Edge[]) => {
+    await localForage.setItem('workflowEdges', edges)
+  }, [])
+
+  const fetchForageWorkflowEdges = useCallback(async () => {
+    var workflowEdges = await localForage.getItem<any>("workflowEdges")
+    if ((!workflowEdges) || (workflowEdges.length === 0)) {
+      workflowEdges = []
+    }
+    return workflowEdges
+  }, [])
+
+  useEffect(() => {
+    (async () => {
+      const forageEdges = await fetchForageWorkflowEdges()
+      await setForageWorkflowEdges(forageEdges)
+      setLoadingEdges(false)
+    })()
+  }, [])
+
+  useEffect(() => {
+    (async () => {
+      if (loadingEdges) {
+        return
+      }
+      await setForageWorkflowEdges(edges)
+    })()
+  }, [edges, setForageWorkflowEdges, loadingEdges])
+
+  const value: IWorkflowsEdgesContext = {
+    edges,
+    fetchForageWorkflowEdges: () => fetchForageWorkflowEdges(),
+    setEdges,
+  }
+
+  return (
+    <WorkflowsEdgesContext.Provider value={value}>
+      {children}
+    </WorkflowsEdgesContext.Provider>
+  )
+}
+
+export default WorkflowsEdgesProvider
