+++ frontend/src/pages/private/workflows/workflows-editor/workflows-editor-page.component.tsx
@@ -1,6 +1,8 @@
 import { PrivateLayout } from 'modules/layout'
+import ProviderContextWrapper from 'context/workflows/workflows-editor.context/provider-context-wrapper'
 import { WorkflowsEditorComponent } from './components/workflows-editor.component'
 import { Grid } from '@mui/material'
+import { WorkflowsEditorProvider } from 'context/workflows/workflows-editor.context'
 /**
  * Workflows editor page
  */
@@ -11,7 +13,11 @@ export const WorkflowsEditorPage = () => {
     <PrivateLayout>
       <Grid container >
         <Grid item xs={12}>
-          <WorkflowsEditorComponent />
+          <ProviderContextWrapper>
+            <WorkflowsEditorProvider>
+              <WorkflowsEditorComponent />
+            </WorkflowsEditorProvider>
+          </ProviderContextWrapper>
         </Grid>
       </Grid>
     </PrivateLayout>
